+++ src/app/Components/ComponentsSamples/Saving/SaveViewSamples.xaml.cs
@@ -1,9 +1,0 @@
-namespace Components.ComponentsSamples.Saving;
-
-public partial class SaveViewSamples 
-{
-    public SaveViewSamples()
-    {
-        InitializeComponent();
-    }
-}