+++ src/app/Playground/HåvardSamples/HåvardPageViewModel.cs
@@ -7,6 +7,9 @@ public class HåvardPageViewModel : ViewModel
 {
     private IEnumerable<object> m_selectedItems;
     private object m_selectedItem;
+    private object m_selectedItem2;
+    private List<Something> m_items;
+    private List<Something> m_items2;
 
     public ICommand Command { get; }
 
@@ -35,18 +38,25 @@ public class HåvardPageViewModel : ViewModel
         SelectedItems = items.ToList();
         SelectedItem = items.First();
 
+        Items2 = items.ToList();
+
         SelectedItemCommand = new Command<IEnumerable<object>>(selectedItems =>
         {
         });
 
         Command = new Command(() =>
         {
-            SelectedItem = null;
+            var oldItems = Items.Take(new Range(0,1));
+            Items = oldItems.ToList();
         });
     }
 
 
-    public List<Something> Items { get; set; }
+    public List<Something> Items
+    {
+        get => m_items;
+        set => RaiseWhenSet(ref m_items, value);
+    }
 
     public IEnumerable<object> SelectedItems
     {
@@ -59,7 +69,26 @@ public class HåvardPageViewModel : ViewModel
     public object? SelectedItem
     {
         get => m_selectedItem;
-        set => RaiseWhenSet(ref m_selectedItem, value);
+        set
+        {
+            Items2 = Items.ToList().Take(new Range(0, 5)).ToList();
+            RaiseWhenSet(ref m_selectedItem, value);
+        }
+    }
+
+    public object SelectedItem2
+    {
+        get => m_selectedItem2;
+        set
+        {
+            RaiseWhenSet(ref m_selectedItem2, value);
+        }
+    }
+
+    public List<Something> Items2
+    {
+        get => m_items2;
+        set => RaiseWhenSet(ref m_items2, value);
     }
 }