+++ src/library/DIPS.Mobile.UI/Components/Chips/Chip.cs
@@ -1,11 +1,17 @@
 using DIPS.Mobile.UI.Extensions;
 using DIPS.Mobile.UI.Resources.Colors;
+using DIPS.Mobile.UI.Resources.Styles.Chip;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.Chips;
 
 public partial class Chip : View
 {
+    public Chip()
+    {
+        Style = InputStyle.Current;
+    }
+
     public void SendTapped()
     {
         Command?.Execute(CommandParameter);