+++ src/library/DIPS.Mobile.UI/Components/Chips/dotnet/ChipHandler.cs
@@ -25,4 +25,14 @@ public partial class ChipHandler : ViewHandler<Chip, ViewHandler>
     {
         throw new Only_Here_For_UnitTests();
     }
+    
+    private static partial void MapBorderColor(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
+
+    private static partial void MapBorderWidth(ChipHandler handler, Chip chip)
+    {
+        throw new Only_Here_For_UnitTests();
+    }
 }