+++ src/library/DIPS.Mobile.UI/Resources/Styles/Chip/ChipStyle.cs
@@ -1,0 +1,8 @@
+namespace DIPS.Mobile.UI.Resources.Styles.Chip;
+
+public enum ChipStyle
+{
+    None=0,
+    Input = 1,
+    EmptyInput = 2,
+}